chore(SBOMER-218): changed packages to better reflect purpose of classes
