+++ service/src/main/java/org/jboss/sbomer/service/rest/ApplicationObjectMapperCustomizer.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.sbomer.service.feature.sbom.rest;
+package org.jboss.sbomer.service.rest;
 
 import jakarta.inject.Singleton;
